// src/frontend-applications/agent-portal/modules/job/model/ExecPlanProperty.ts
/**
 * Copyright (C) 2006-2025 KIX Service Software GmbH, https://www.kixdesk.com
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum ExecPlanProperty {

    ID = 'ID',

    NAME = 'Name',

    TYPE = 'Type',

    PARAMETERS = 'Parameters'

}